搜尋ok

+++ themes/branch_list.php
@@ -50,14 +50,10 @@ if ($totalRows) {
         <!-- 分頁膠囊 -->
         <ul class="nav nav-pills mb-4" id="pills-tab" role="tablist">
           <li class="nav-item me-3" role="presentation">
-            <button class="nav-link active rounded-pill fw-bold" id="pills-home-tab" data-bs-toggle="pill"
-              data-bs-target="#pills-home" type="button" role="tab" aria-controls="pills-home"
-              aria-selected="true">分店列表</button>
+            <button class="nav-link active rounded-pill fw-bold" id="pills-home-tab" data-bs-toggle="pill" data-bs-target="#pills-home" type="button" role="tab" aria-controls="pills-home" aria-selected="true">分店列表</button>
           </li>
           <li class="nav-item" role="presentation">
-            <button class="nav-link rounded-pill fw-bold" id="pills-profile-tab" data-bs-toggle="pill"
-              data-bs-target="#pills-profile" type="button" role="tab" aria-controls="pills-profile"
-              aria-selected="false">新增分店</button>
+            <button class="nav-link rounded-pill fw-bold" id="pills-profile-tab" data-bs-toggle="pill" data-bs-target="#pills-profile" type="button" role="tab" aria-controls="pills-profile" aria-selected="false">新增分店</button>
           </li>
           <li class="ms-auto">
 
@@ -98,7 +94,7 @@ if ($totalRows) {
                 </tr>
               </thead>
               <tbody>
-                <?php foreach ($rows as $r): ?>
+                <?php foreach ($rows as $r) : ?>
                   <tr>
                     <td><?= $r['id'] ?></td>
                     <td><?= $r['branch_name'] ?></td>
@@ -107,18 +103,14 @@ if ($totalRows) {
                     <td><?= $r['open_time'] ?></td>
                     <td><?= $r['close_time'] ?></td>
                     <td><?= $r['branch_status'] ?></td>
-                    <td>
-                      <a href="branch_reservation_details.php?id=<?= $r['id'] ?>">
-                        <i class="fa-solid fa-user-check"> 預約</i></i>
-                      </a>
-                    </td>
+                    <td><a href="branch_reservation_details.php?id=<?= $r['id'] ?>">
+                        <i class="fa-solid fa-user-check"> 預約</i></a></td>
                     <td>
                       <a href="branch_edit.php?id=<?= $r['id'] ?>">
                         <i class="fa-solid fa-pen-to-square"></i>
                       </a>
                     </td>
-                    <td><a href="branch_delete.php?id=<?= $r['id'] ?>"
-                        onclick="return confirm('是否要刪除編號為<?= $r['id'] ?>的資料')">
+                    <td><a href="branch_delete.php?id=<?= $r['id'] ?>" onclick="return confirm('是否要刪除編號為<?= $r['id'] ?>的資料')">
                         <i class="fa-solid fa-trash text-danger"></i>
                       </a></td>
                   </tr>
@@ -139,12 +131,12 @@ if ($totalRows) {
                       <i class="fa-solid fa-angle-left"></i>
                     </a>
                   </li>
-                  <?php for ($i = $page - 5; $i <= $page + 5; $i++):
-                    if ($i >= 1 and $i <= $totalPages): ?>
+                  <?php for ($i = $page - 5; $i <= $page + 5; $i++) :
+                    if ($i >= 1 and $i <= $totalPages) : ?>
                       <li class="page-item <?= $page == $i ? 'active' : '' ?> ">
                         <a class="page-link" href="?page=<?= $i ?>"><?= $i ?></a>
                       </li>
-                    <?php endif;
+                  <?php endif;
                   endfor ?>
                   <li class="page-item">
                     <a class="page-link" href="#">
@@ -174,13 +166,13 @@ if ($totalRows) {
 <?php include __DIR__ . '/../parts/scripts.php' ?>
 
 <script>
-  document.getElementById('searchForm').addEventListener('submit', function (event) {
-    event.preventDefault(); // 阻止表单提交
+  document.getElementById('searchForm').addEventListener('submit', function(event) {
+    event.preventDefault(); // 阻止表單提交
 
     var formData = new FormData(this);
-    var queryString = new URLSearchParams(formData).toString(); // 将表单数据转换为 URL 查询字符串
+    var queryString = new URLSearchParams(formData).toString(); // 將表單數據轉換為 URL 查詢字符串
 
-    fetch('branch_list_search.php?' + queryString) // 将查询字符串附加到 URL 中
+    fetch('branch_list_search.php?' + queryString) // 將查詢字符串附加到 URL 中
       .then(response => response.json())
       .then(data => {
         var tableBody = document.createElement('tbody');
@@ -200,16 +192,11 @@ if ($totalRows) {
         `;
           tableBody.appendChild(row);
         });
-        // 更新表格内容
-        var branchListTable = document.getElementById('branchListTable');
-        branchListTable.innerHTML = '';
-        branchListTable.appendChild(tableBody);
-      })
-      .catch(error => {
-        console.error('Error:', error);
+        var table = document.getElementById('branchListTable');
+        table.querySelector('tbody').remove(); // 移除先前的 tbody 元素
+        table.appendChild(tableBody); // 添加新的 tbody 元素
       });
   });
-
 </script>
 
 <?php include __DIR__ . '/../parts/html-foot.php' ?>
